Changed binding got LineAlignCommand to use &,m_lineAlignCommand
While this doesn't make a lot of difference for a "WhenHeld" operation (constructors for use with either type of reference exist and work fine in this case), it becomes more of an issue when using other types of bindings.

A key consideration that comes into pay when using pairs of bindings like "WhenPressed" to start a command and "CancelWhenPressed" to cancel a command *must* be using the same instance of a command.  Using & here ensures that are referring to the same instance of a command when we start start it and when we cancel it.  You can run multiple copies of the same command at the same time and when you do a Cancel on the command ("CancelWhenPressed" does a command->Cancel() using the passed command), it is possibly that you could end up calling Cancel on a command instance which is not running - this isn't an error and the scheduler handles it as expected.

But what happens is you don't "Cancel" the instance of the command you think you are and the entire operation becomes a no-op of cancelling a command that wasn't running anyway (or that the scheduler thought wasn't running).

Using & of an existing command instance ensures that the "Cancel" is being done on the same command that was started and is a better and more reliable practice.